+++ euler579c/cube.h
@@ -9,7 +9,7 @@ class cube
 {
 public:
 	static long long maxSize;
-	int width, height, depth;
+	long long width, height, depth;
 	set<vertex> vertices;
 	cube(const vector3d& U, const vector3d& V, const vector3d& N, bool flipX, bool flipY, bool flipZ, int* order);
 	bool is_oversize() const;
