+++ euler579c/vector3d.h
@@ -21,6 +21,6 @@ public:
 };
 
 vertex operator+(const vertex&, const vector3d&);
-vector3d operator/(const vector3d&, long);
+vector3d operator/(const vector3d&, long long);
 
 #endif
